Debug Pile

+++ src_asserv/include/config.h
@@ -20,6 +20,21 @@
 #include "../hardware/i2c.h"
 
 #include "struct_Commande.h"
+#include "Lecture_trapeze.h"
+
+#include "Calcule_distance.h"
+#include "detecBugBloquage.h"
+
+#include "Modulo_pi.h"
+#include "Mouvement_Robot.h"
+#include "Nombre_ordre_enregistre.h"
+#include "Odometrie.h"
+#include "Ordres.h"
+#include "PI_Asservissement.h"
+#include "Pile.h"
+#include "Ralentissement_free.h"
+#include "STOP_Force.h"
+#include "traitementI2C.h"
 
 
 //////////////////////////////////////////////////////////////////////////
@@ -34,11 +49,11 @@
 #define Sens_codeur_G	1
 
 //PI
-#define Ki_U 0//0.2
-#define Kp_U 1//0.8
+#define Ki_U (double)0.0//0.2
+#define Kp_U (double)1.0//0.8
 
-#define Ki_T 5//0.005
-#define Kp_T 0//1.2
+#define Ki_T (double)5.0//0.005
+#define Kp_T (double)0.0//1.2
 
 //Commande moteur
 #define Sens_mot_D -1
